PDCL-7517 - move createDataCollectionRequest from core to utils/request
