+++ src/utils/request/createDataCollectionRequest.js
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { createRequest } from "../../../utils/request";
+import createRequest from "./createRequest";
 
 export default dataCollectionRequestPayload => {
   const getUseSendBeacon = ({ isIdentityEstablished }) => {
